+++ e2e/e2etests/test_zeta_withdraw_bitcoin_revert.go
@@ -4,7 +4,7 @@ import (
 	"math/big"
 
 	"github.com/stretchr/testify/require"
-	connectorzevm "github.com/zeta-chain/protocol-contracts/pkg/contracts/zevm/zetaconnectorzevm.sol"
+	connectorzevm "github.com/zeta-chain/protocol-contracts/v1/pkg/contracts/zevm/zetaconnectorzevm.sol"
 
 	"github.com/zeta-chain/zetacore/e2e/runner"
 	"github.com/zeta-chain/zetacore/e2e/utils"