+++ x/crosschain/keeper/grpc_query_inbound_hash_to_cctx_test.go
@@ -135,6 +135,7 @@ func createInTxHashToCctxWithCctxs(keeper *crosschainkeeper.Keeper, ctx sdk.Cont
 		cctxs[i].ZetaFees = math.OneUint()
 		cctxs[i].InboundParams = &types.InboundParams{ObservedHash: fmt.Sprintf("%d", i), Amount: math.OneUint()}
 		cctxs[i].CctxStatus = &types.Status{Status: types.CctxStatus_PendingInbound}
+		cctxs[i].RevertOptions = types.NewEmptyRevertOptions()
 		keeper.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, cctxs[i])
 	}